% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathR.R
\name{bathR}
\alias{bathR}
\title{I am a bathR}
\usage{
bathR(bathRs = 1)
}
\arguments{
\item{bathRs}{double The ASCII image you wish to print to the console}
}
\value{
Some ASCI-bathing love
}
\description{
I am a bathR
}
\examples{
bathR()
}
